Move animation and core tests to Jest
