+++ .gitignore
@@ -1,5 +1,6 @@
-node_modules/*
-build/*
+node_modules/
+coverage/
+build/
 *.tgz
 *.log
 *.swp
