+++ package.json
@@ -20,6 +20,7 @@
     "rollup_dev": "rollup -c -w"
   },
   "devDependencies": {
+    "canvas": "^2.3.1",
     "eslint": "5.3.0",
     "eslint-config-airbnb": "17.1.0",
     "eslint-plugin-import": "^2.14.0",
@@ -35,6 +36,7 @@
     "http-server": "~0.9.0",
     "husky": "^1.3.1",
     "jest": "^24.1.0",
+    "jest-canvas-mock": "^2.0.0-alpha.3",
     "lint-staged": "^8.1.4",
     "load-grunt-tasks": "~0.6.0",
     "prettier": "^1.16.4",
@@ -87,6 +89,11 @@
   "babel": {
     "presets": [
       "@babel/preset-env"
+    ]
+  },
+  "jest": {
+    "setupFiles": [
+      "jest-canvas-mock"
     ]
   },
   "dependencies": {
